Create Jeff Jampol
Ok Jeff Jampol was somehow managing the doors and my Jim Morrison and the Morrison estate for which I Jim Morrison Jr am the Air to. Jeff Jampol is a criminal and a drug addict and was a homeless man before some how gaining access to my dad's estate . I need help he has hired IT Techs to block my information and put redirects on all money links that have anything to do with linking me to my father or the doors or anything positive towards my endeavors.